v20200331

+++ drivers/net/wireless/rtl8189fs/os_dep/linux/ioctl_cfg80211.c
@@ -348,7 +348,7 @@ rtw_cfg80211_default_mgmt_stypes[NUM_NL80211_IFTYPES] = {
 static u64 rtw_get_systime_us(void)
 {
 #if (LINUX_VERSION_CODE >= KERNEL_VERSION(4,20,0))
-	return ktime_to_us(ktime_get_boottime());
+	return (u64)ktime_to_us(ktime_get_boottime());
 #elif (LINUX_VERSION_CODE >= KERNEL_VERSION(2,6,39))
 	struct timespec ts;
 	get_monotonic_boottime(&ts);
